refactor: http_file_transfer
