+++ .gitignore
@@ -8,7 +8,8 @@ conan-cache/
 coverage.xml
 
 # User specific settings
-
+compile_commands.json
+CMakeUserPresets.json
 
 # IDE files
 .vs/
@@ -66,4 +67,4 @@ $RECYCLE.BIN/
 .TemporaryItems
 ehthumbs.db
 Thumbs.db
-compile_commands.json
+
