+++ core/http_file_transfer/private/include/constants.hpp
@@ -1,0 +1,16 @@
+
+#ifndef CPP_CORE_PROJECT_CONSTANTS_HPP
+#define CPP_CORE_PROJECT_CONSTANTS_HPP
+
+#define APP_JSON "application/json"
+#define TEXT_PLAIN "text/plain"
+#define OCTET_STREAM "application/octet-stream"
+#define CONTENT_TYPE "Content-Type"
+#define CACHE_CONTROL "Cache-Control"
+#define CONTENT_RANGE "Content-Range"
+#define CONTENT_LENGTH "Content-Length"
+#define CONTENT_DISPOSITION "Content-Disposition"
+
+#define CHUNK_SIZE (32 * 1024)
+
+#endif // CPP_CORE_PROJECT_CONSTANTS_HPP
