+++ core/http_file_transfer/private/include/helpers.hpp
