+++ core/http_file_transfer/private/src/mylib.cpp
@@ -1,0 +1,7 @@
+#include "mylib.h"
+
+// ===========================================================
+// MyLib
+// ===========================================================
+
+int Sum(const int pLhs, const int pRhs) { return pLhs + pRhs; }
